Create a new channel cross section

Create a channel cross section by specifying its properties.

Usage

cross_section(width, grad, d50, d84, roughness, rootdepth = 0)

Arguments

width

Channel width in meters; single positive numeric.

grad

Energy gradient of the stream channel; single positive numeric.

d50, d84

Grain size distribution's 50th and 84th quantiles in millimeters; single positive numerics.

roughness

Manning's roughness; positive single numeric.

rootdepth

Effective rooting depth for vegetation; single non-negative numeric.

Value

A "cross_section" object.

Details

Examples of effective rooting depth for vegetation, rootdepth, are:

grassy banks, no trees / shrubs: rootdepth = 0.35.

1 to 5% tree / shrub cover: rootdepth = 0.50.

5 to 50% tree / shrub cover: rootdepth = 0.90.

more than 50% tree / shrub cover: rootdepth = 1.10.

Note

Width is assumed constant as a function of depth.

Examples

cross_section(3, grad = 0.01, d50 = 45, d84 = 80, roughness = 0.01)
#> Channel with width 3